requirements.txt
opencv-python
numpy
Pillow
torch
torchvision
pycocotools
websockets
